refactor(redux): вынести редьюсер в отдельный файл
